mbl_cloneables_factory: split into def and implementaiton
The use of the MBL_CLONEABLES_FACTORY_INSTANTIATE() macro after the
implementation of mbl_cloneables_factory was causing the compilation error
"error: explicit specialization of 'objects' after instantiation". Breaking
the definition and implementation into two files fixes this issue.
