+++ contrib/mul/clsfy/Templates/mbl_cloneables_factory+clsfy_builder_1d-.cxx
@@ -1,3 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <clsfy/clsfy_builder_1d.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(clsfy_builder_1d);
+#include <mbl/mbl_cloneables_factory.h>
