+++ contrib/mul/mbl/tests/test_cloneables_factory.cxx
@@ -1,5 +1,5 @@
 // This is mul/mbl/tests/test_cloneables_factory.cxx
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <testlib/testlib_test.h>
 
 
@@ -29,6 +29,7 @@ class mbl_test_cf_B : public mbl_test_cf_base
 
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mbl_test_cf_base);
 
+#include <mbl/mbl_cloneables_factory.h>
 
 void test_cloneables_factory()
 {
