+++ contrib/mul/mcal/Templates/mbl_cloneables_factory+mcal_component_analyzer-.cxx
@@ -1,3 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mcal/mcal_component_analyzer.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mcal_component_analyzer);
+#include <mbl/mbl_cloneables_factory.h>
