+++ contrib/mul/mcal/Templates/mbl_cloneables_factory+mcal_single_basis_cost-.cxx
@@ -1,3 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mcal/mcal_single_basis_cost.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mcal_single_basis_cost);
+#include <mbl/mbl_cloneables_factory.h>
