+++ contrib/mul/mfpf/Templates/mbl_cloneables_factory+mfpf_point_finder-.cxx
@@ -1,3 +1,4 @@
 #include <mfpf/mfpf_point_finder.h>
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mfpf_point_finder);
+#include <mbl/mbl_cloneables_factory.h>
