+++ contrib/mul/mfpf/Templates/mbl_cloneables_factory+mfpf_point_finder_builder-.cxx
@@ -1,4 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mfpf/mfpf_point_finder_builder.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mfpf_point_finder_builder);
-
+#include <mbl/mbl_cloneables_factory.h>
