+++ contrib/mul/mfpf/Templates/mbl_cloneables_factory+mfpf_region_definer-.cxx
@@ -1,4 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mfpf/mfpf_region_definer.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mfpf_region_definer);
-
+#include <mbl/mbl_cloneables_factory.h>
