+++ contrib/mul/mfpf/Templates/mbl_cloneables_factory+mfpf_vec_cost_builder-.cxx
@@ -1,4 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mfpf/mfpf_vec_cost_builder.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mfpf_vec_cost_builder);
-
+#include <mbl/mbl_cloneables_factory.h>
