+++ contrib/mul/mipa/Templates/mbl_cloneables_factory+mipa_vector_normaliser-.cxx
@@ -1,3 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mipa/mipa_vector_normaliser.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mipa_vector_normaliser);
+#include <mbl/mbl_cloneables_factory.h>
