+++ contrib/mul/mmn/Templates/mbl_cloneables_factory+mmn_solver-.cxx
@@ -1,4 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <mmn/mmn_solver.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(mmn_solver);
-
+#include <mbl/mbl_cloneables_factory.h>
