+++ contrib/mul/msm/Templates/mbl_cloneables_factory+msm_aligner-.cxx
@@ -1,3 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <msm/msm_aligner.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(msm_aligner);
+#include <mbl/mbl_cloneables_factory.h>
