+++ contrib/mul/msm/Templates/mbl_cloneables_factory+msm_param_limiter-.cxx
@@ -1,3 +1,4 @@
-#include <mbl/mbl_cloneables_factory.h>
+#include <mbl/mbl_cloneables_factory_def.h>
 #include <msm/msm_param_limiter.h>
 MBL_CLONEABLES_FACTORY_INSTANTIATE(msm_param_limiter);
+#include <mbl/mbl_cloneables_factory.h>
